requirements.txt
langchain
langchain-community
langchain-openai
langchain-text-splitters
python-dotenv
chromadb
sentence-transformers
pypdf
python-docx
flashrank
rank_bm25
flask
flask-cors
